+++ controller/photoController.js
@@ -1,18 +1,16 @@
 const {Photo, User, Comment} = require('./../models')
 
 
-
 class PhotoController {
 
     static async addPhoto(req, res){
-
         const UserId = req.users.id
         const {poster_image_url, title, caption} = req.body
         const data =  {poster_image_url, title, caption, UserId}
         try {
             const photos = await Photo.create(data)
             if (!photos) throw { name: 'SequelizeValidationError' }
-            res.status(201).json(photos)
+            res.status(201).json({photo:photos})
         } catch (error) {
             if (error.name === 'SequelizeValidationError') {
                 const ValidationError = error.errors.map((error) => {
@@ -42,11 +40,11 @@ class PhotoController {
                 }]
             })
             if(!photos) throw {name:"DataNotFound"}
-            res.status(200).json(photos)
+            res.status(200).json({photos})
         } catch (error) {
             if (error.name === "DataNotFound") {
                 res.status(404).json("Data Not Found")
-            } else {
+            }  else {
                 res.status(500).json(error)
             }
         }
@@ -59,12 +57,19 @@ class PhotoController {
         const {title, caption, poster_image_url} = req.body 
         const data = {title, caption, poster_image_url}
         try {
-            const photos = await Photo.update(data, {where:{id}, attributes:['id','title','caption','poster_image_url'], returning: true}) 
-            if(photos[0]===0) throw ({name:"cantDelete"})
-            res.status(200).json(photos[1])
+            const photos = await Photo.update(data, {where:{id, UserId:req.users.id}, returning: true}) 
+            if(photos[0]===0) throw ({name:"cantUpdatePhoto"})
+            if (!photos) throw { name: 'SequelizeValidationError' }
+            const datas = photos[1][0]
+            res.status(200).json({photo:datas})
         } catch (error) {
-            if(error.name === "cantDelete"){
-                res.status(400).json(error)
+            if(error.name === "cantUpdatePhoto" ){
+                res.status(400).json({message:`cant Update photos where photoId : ${id} or body request is null`})
+            } else if (error.name === 'SequelizeValidationError') {
+                const ValidationError = error.errors.map((error) => {
+                    return error.message
+                })
+                res.status(400).json({ message: ValidationError })
             } else{
                 res.status(500).json(error)
             }
@@ -74,14 +79,14 @@ class PhotoController {
     static async removePhoto(req,res){ 
         const {photoId} = req.params
         const id = photoId
-        console.log(id);
+        
         try {
-            const photos = await Photo.destroy({ where: {id} })
+            const photos = await Photo.destroy({ where: {id, UserId:req.users.id} })
             if (!photos) throw { name: 'ErrNotFound' }
             res.status(200).json({message:"Your photos has been succesfully deleted"})
         } catch (error) {
-            if (error.name === 'ErrNotFound') {
-                res.status(404).json({ message: 'data not found' })
+            if (error.name === 'ErrNotFound' || error.name === 'SequelizeDatabaseError') {
+                res.status(404).json({ message: 'photoId not found' })
             } else {
                 res.status(500).json(error.message)
             }
